Change price to int

+++ src/main/java/org/example/Coffee.java
@@ -1,6 +1,6 @@
 package org.example;
 
 public interface Coffee {
-    double getPrice();
+    int getPrice();
     String getDescription();
 }
