+++ src/main/java/org/example/CreditCardStrategy.java
@@ -2,7 +2,7 @@ package org.example;
 
 public class CreditCardStrategy implements PaymentStrategy {
     @Override
-    public String checkout(double amount) {
+    public String checkout(int amount) {
         return "Checkout with Credit Card: $" + amount;
     }
 }
