+++ src/main/java/org/example/PaypalStrategy.java
@@ -2,7 +2,7 @@ package org.example;
 
 public class PaypalStrategy implements PaymentStrategy {
     @Override
-    public String checkout(double amount) {
+    public String checkout(int amount) {
         return "Checkout with Paypal: $" + amount;
     }
 }
